Lower RAM usage for tiles
This lowers RAM usage by an extreme amount. On windows, roughly 1/2 of
the RAM is used now, and, with mono, roughly 1/4-1/5 of the RAM is used
now, bringing servers from 1GB+ down to a manageable 350MB for a large
world.

Unfortunately, this breaks compatibility with all existing tile-based
plugins, but the gains are more than the losses of having to recode the
existing tile-based plugins.